requirements.txt
Flask
requests
python-dotenv
pyotp
protobuf
googleapis-common-protos